backend/requirements.txt
Django==4.2.18
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
PyJWT==2.6.0
python-dotenv==1.0.1
Pillow==10.2.0
openai==1.20.1
cryptography==39.0.0
pyotp==2.9.0
pytz==2024.1 